mcp_a2a/pyproject.toml
[project]
name = "mcp-a2a"
version = "0.1.0"
description = "A simple Starlette web application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.0",
    "google-adk>=1.2.0",
    "starlette>=0.37.0",
    "uvicorn[standard]>=0.24.0",
]
